+++ AmazonChimeSDK/AmazonChimeSDK/audiovideo/video/capture/DefaultCameraCaptureSource.swift
@@ -56,6 +56,7 @@ import UIKit
         NotificationCenter.default.removeObserver(self)
     }
 
+    // TODO: `device` is being initialize twice: here and in init(), assess if this is needed
     public var device: MediaDevice? = MediaDevice.listVideoDevices().first {
         didSet {
             guard let device = device else { return }
@@ -189,6 +190,11 @@ import UIKit
 
         if device != nil {
             eventAnalyticsController?.pushHistory(historyEventName: .videoInputSelected)
+        }else {
+            let attributes = [
+                EventAttributeName.videoInputError: DeviceError.noCameraSelected
+            ]
+            eventAnalyticsController?.publishEvent(name: .videoInputFailed, attributes: attributes)
         }
     }
 
@@ -259,10 +265,9 @@ import UIKit
 
     private func handleCaptureFailed(reason: CaptureSourceError) {
         let attributes = [
-            EventAttributeName.videoInputError: reason
+            EventAttributeName.deviceAccessFailedError: reason
         ]
-
-        eventAnalyticsController?.publishEvent(name: .videoInputFailed, attributes: attributes)
+        eventAnalyticsController?.publishEvent(name: .deviceAccessFailed, attributes: attributes)
 
         ObserverUtils.forEach(observers: captureSourceObservers) { (observer: CaptureSourceObserver) in
             observer.captureDidFail(error: reason)