+++ AmazonChimeSDK/AmazonChimeSDK/ingestion/IngestionPayload.swift
@@ -22,6 +22,7 @@ import Foundation
     public let retryCount: Int?
     public let videoInputErrorMessage: String?
     public let audioInputErrorMessage: String?
+    public let deviceAccessErrorMessage: String?
     public let ttl: Int64?
 
     public init(name: String,
@@ -37,6 +38,7 @@ import Foundation
                 retryCount: Int? = nil,
                 videoInputErrorMessage: String? = nil,
                 audioInputErrorMessage: String? = nil,
+                deviceAccessErrorMessage: String? = nil,
                 ttl: Int64? = nil) {
         self.name = name
         self.ts = ts
@@ -51,6 +53,7 @@ import Foundation
         self.retryCount = retryCount
         self.videoInputErrorMessage = videoInputErrorMessage
         self.audioInputErrorMessage = audioInputErrorMessage
+        self.deviceAccessErrorMessage = deviceAccessErrorMessage
         self.ttl = ttl
     }
 }